update time range
